Forgot the PIN
Once a PIN is added to an account, it will be needed to make changes to the account settings.

At this time, users and their parents/guardians will not be able to manually reset the PIN if it is forgotten. Customer Support will need to assist with resetting forgotten PINs. To reach Customer Support, please use the contact form. 

 

Note: Customer Support may request specific information to ensure that they are speaking to the parent/guardian before a PIN is reset.